refactor(primitives): extract ByteReader and ConstantTime into Primitives
ByteReader previously lived at file scope in
Auth/certificates/Certificate.swift and was documented as
"internal to Certificate". In practice it is a generic cursor-based
reader over Data that any byte-parsing code in the SDK could reuse,
so it moves to Sources/BSV/Primitives/ByteReader.swift. The existing
Transaction parser uses a plain inout Int offset and is a different
abstraction, so the two are kept separate.

ConstantTime was defined at file scope in Wallet/ProtoWallet.swift
but already reached across modules — AuthNonce.swift depends on it
for HMAC verification. That cross-module reach is the wrong shape:
Auth should not need to know that a constant-time comparator lives
inside Wallet. Move it to Sources/BSV/Primitives/ConstantTime.swift
where both Wallet and Auth can depend on it cleanly.

No behavioural changes; all 497 tests continue to pass.

Addresses code review finding M5.
